requirements.txt
anyio==4.0.0
astor==0.8.1
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.1
cycler==0.12.0
decorator==5.1.1
exceptiongroup==1.1.3
fonttools==4.43.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-resources==6.1.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.7.3
mpmath==1.3.0
numpy==1.24.4
onnx==1.14.1
opt-einsum==3.3.0
packaging==23.1
paddle-bfloat==0.1.7
paddlelite==2.13rc0
paddlepaddle==2.4.1
paddleslim==2.4.1
Pillow==10.0.1
protobuf==3.20.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
swig==4.1.1
sympy==1.12
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
x2paddle==1.4.1
zipp==3.17.0
